requirements.txt
pygithub
pygit2